Added default message digest for openssl < 1.0.0
Openssl versions prior to 1.0.0 were giving an error because they didn't have a default message digest type. I added one: sha1

I believe this is how it was fixed before, but they provided md5 as the default digest type and Apple was not cool with that, causing errors when running SiriProxy.