Sort of fixed race condition for dns server
The DNS server runs on a different thread from the proxy server. If the DNS server doesn't start before the proxy server de-privileges itself, the DNS server can't open port 53. So I added in a wait on the main thread of up to 63 seconds. Not the best solution, but it works.